Move JsonWebToken out of Symfony namespace
