+++ app/CommandHandling/ContextFactory.php
@@ -7,7 +7,7 @@ namespace CultuurNet\UDB3\Silex\CommandHandling;
 use Broadway\Domain\Metadata;
 use CultuurNet\UDB3\ApiGuard\ApiKey\ApiKey;
 use CultuurNet\UDB3\ApiGuard\Consumer\Consumer;
-use CultuurNet\UDB3\Jwt\Symfony\Authentication\JsonWebToken;
+use CultuurNet\UDB3\Jwt\JsonWebToken;
 use CultuurNet\UDB3\User\CurrentUser;
 use Silex\Application;
 use Symfony\Component\HttpFoundation\Request;
