+++ src/Jwt/JsonWebToken.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Jwt\Symfony\Authentication;
+namespace CultuurNet\UDB3\Jwt;
 
 use CultuurNet\UDB3\User\UserIdentityDetails;
 use CultuurNet\UDB3\User\UserIdentityResolver;
